refactorizacion del codigo
