+++ firmware/src/html.cpp
@@ -1,13 +1,12 @@
 /*
- *  html.h
- *  Captive Portal Embebido
- *  created 29 Ene 2024
+ *  html.cpp
+ *  Páginas embebidas
+ *  created 31 Ene 2024
  *  Lucas Martin Treser (github: @lmtreser)
  *
 */
 
-#ifndef html_h
-#define html_h
+#include "espWebCfg.h"
 
 const char* htmlConfig PROGMEM = R"(
   <!DOCTYPE html>
@@ -15,7 +14,7 @@ const char* htmlConfig PROGMEM = R"(
     <head>
       <meta charset="UTF-8"/>
       <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-      <title>ESP8266 WebServer Config</title>
+      <title>ESP WebServer Config</title>
       <style>
         /* Estilos generales del formulario */
         form {  
@@ -57,7 +56,7 @@ const char* htmlConfig PROGMEM = R"(
       </style>
     </head>
     <body>
-      <h1>Simple Portal Web para ESP8266</h1>
+      <h1>Simple Portal Web para ESP8266/ESP32</h1>
       <p>Ingresar los datos de la red WiFi a la que se desea conectar:</p>
       <form action="/" method="post">
         <input type="text" name="ssid" maxlength="30" placeholder="SSID"/><br/>
@@ -99,5 +98,3 @@ const char* htmlSuccess PROGMEM = R"(
       </div>
       </body>
   </html>)";
-
-#endif
